+++ SchoolManagement/SchoolManagementAPI/SchoolManagementAPI/Controllers/V1/CourseController.cs
@@ -93,6 +93,11 @@ namespace SchoolManagementAPI.Controllers.V1
                 _logger.LogError(new ArgumentOutOfRangeException(nameof(id)), "Id field can't be <= zero OR it doesn't match with model's Id.");
                 return BadRequest();
             }
+            var available = await _courseService.GetCourseName(courseVm.CourseName, courseVm.GradeId);
+            if (available != null)
+            {
+                return BadRequest("Course already exist");
+            }
             var data = _mapper.Map<CourseVm, Course>(courseVm);
             var result = await _courseService.UpdateCourseAsync(id, data);
             if (result is null)
@@ -115,10 +120,15 @@ namespace SchoolManagementAPI.Controllers.V1
                 _logger.LogError(new ArgumentOutOfRangeException(nameof(id)), "Id field can't be {id}", id);
                 return BadRequest();
             }
-            var result = await _courseService.DeleteAsync(id);
-            if (result is null)
-                return NotFound();
-            return Ok(result);
+            var existingData=await _courseService.GetCourseAsync(id);
+            if (existingData != null)
+            {
+                var result = await _courseService.DeleteAsync(id);
+                if (result is null)
+                    return NotFound();
+                return Ok(result);
+            }
+            return BadRequest("Course not found");
         }
 
     }
